Support the manufacturing version query for OSInfo
:Release Notes:
nyx-cmd can now be used to query the operating system's manufacturing
version.

:Detailed Notes:
nyx-lib 5.2.0 now provides, via the OSInfo module, a value for the
manufacturing version of the operating system. The OSInfo module
is extended to support this field as a command line argument.

Also, refactored an error reporting block to eliminate a desktop
compilation error.

:Testing Performed:
Compiled locally and tested on desktop and emulator.

:QA Notes:

:Issues Addressed:
[GF-12714] Supply a "manufacturing version" for webOS
[GF-13455] nyx-cmd: Add support for NYX_OS_INFO_MANUFACTURING_VERSION query

Change-Id: I8b4ca891fb76da6bdc38483db6f738eb68485b48
Reviewed-on: https://gpro.palm.com/9338
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Sapna Todwal <sapna.todwal@lge.com>