(Finally) Release nyx-utils 1.0.0 to the public
:Release Notes:
nyx-utils is published on github as an open source component, with a
version of 1.0.0.

:Detailed Notes:
The recipe for nyx-utils has always resided in meta-webos, but the
actual source code was not approved for release until recently.

This commit takes the component to an official 1.0.0 version.

All copyrights are assigned to LG Electronics.

:Testing Performed:
This is a build time change only. Code changes affect only comments
other than finalizing the version number.

:QA Notes:

:Issues Addressed:
[GF-5644]  Convert HP copyright to LG
[GF-8118]  nyx-utils recipe fails to build in build-webos
[GF-32659] Release Nyx-utils to the public

Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>

Change-Id: Ifcfc5b530d51e8d43dedce007bf76b5ca2906a41
Reviewed-on: https://g2g.palm.com/3195
Reviewed-by: Build Verification
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>